Change nRefHeight in version=2 transactions to be 32-bit instead of 64-bit. Fixes GH #56, but requires new genesis block.

+++ src/main.h
@@ -401,7 +401,7 @@ public:
     std::vector<CTxIn> vin;
     std::vector<CTxOut> vout;
     unsigned int nLockTime;
-    int64 nRefHeight;
+    int32_t nRefHeight;
 
     // Denial-of-service detection:
     mutable int nDoS;
